+++ Gruntfile.js
@@ -224,44 +224,6 @@ module.exports = function(grunt) {
     },
 
     /**
-     * Copy: https://github.com/gruntjs/grunt-contrib-copy
-     *
-     * Copy files and folders.
-     */
-    copy: {
-      'static-legacy': {
-        files:
-        [
-          {
-            expand: true,
-            cwd: 'vendor/cf-core/',
-            src: [
-              'licensed-fonts.css'
-            ],
-            dest: 'static-legacy/css/'
-          }
-        ]
-      },
-      vendor: {
-        files:
-        [
-          {
-            expand: true,
-            cwd: '',
-            src: [
-              // Only include vendor files that we use independently
-              'vendor/html5shiv/html5shiv-printshiv.min.js',
-              'vendor/box-sizing-polyfill/boxsizing.htc',
-              'vendor/slick-carousel/slick.min.js',
-              'vendor/slick-carousel/slick.css'
-            ],
-            dest: 'static'
-          }
-        ]
-      }
-    },
-
-    /**
      * Lint the JavaScript.
      */
     lintjs: {
@@ -396,8 +358,7 @@ module.exports = function(grunt) {
   /**
    * Create custom task aliases and combinations.
    */
-  grunt.registerTask('vendor', ['bower:install', 'string-replace:chosen',
-                                'copy:static-legacy', 'concat:cf-less']);
+  grunt.registerTask('vendor', ['bower:install', 'string-replace:chosen', 'concat:cf-less']);
   grunt.registerTask('css', ['less', 'autoprefixer', 'legacssy', 'usebanner:css']);
   grunt.registerTask('js', ['browserify:build', 'usebanner:js']);
   grunt.registerTask('test', ['lintjs', 'mocha_istanbul']);
@@ -406,6 +367,6 @@ module.exports = function(grunt) {
     grunt.task.run(this.target);
   });
 
-  grunt.registerTask('build', ['test', 'css', 'js', 'copy:vendor']);
+  grunt.registerTask('build', ['test', 'css', 'js']);
   grunt.registerTask('default', ['build']);
 };