+++ CHANGELOG.md
@@ -44,6 +44,8 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Office folder and files for Office pages
 - Updated template for office pages
 - Relaxed ESLint `key-spacing` rule to warning.
+- Refactored breakpoint-handler module into separate class modules
+  and utility function.
 
 ### Fixed
 - Fixed macro on offices page template
@@ -60,6 +62,7 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - alert macro
 - Unused index.html file from /initiatives/
 - Unnecessary setting of template variables
+- Removes `copy:static-legacy` and `grunt-contrib-copy` package.
 
 
 ## 3.0.0-1.0.1 - 2015-05-18